[train][data] Enable Train v2 for ray/air/tests (#57534)
* Enables V2 env variable for most `ray/air/tests` and
`py_doctest[air]`.
* Explicitly disables V2 for certain tests that test legacy behavior
such as the old Tune integration, or marks them for migration in a
followup.
* Removes some unmaintained/deprecated code (`DummyTrainer`,
`custom_trainer`)
* Moved some tests to `ray/data` for utils that live in `ray.air` but
are only used in Ray Data. Note that I didn't move the util source code
to Data as that would have been a large change, but I moved the tests
out of this deprecated "air tests" suite:
    * `test_arrow` -> `test_arrow_type_conversion`
    * `test_data_batch_conversion`
    * `test_object_extension`
    * `test_tensor_extension`
    * `test_torch_tensor_utils`

---------
