+++ python/ray/data/BUILD.bazel
@@ -207,6 +207,22 @@ py_test(
 )
 
 py_test(
+    name = "test_arrow_type_conversion",
+    size = "small",
+    srcs = [
+        "tests/test_arrow_type_conversion.py",
+    ],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
     name = "test_arrow_serialization",
     size = "small",
     srcs = ["tests/test_arrow_serialization.py"],
@@ -589,6 +605,20 @@ py_test(
 )
 
 py_test(
+    name = "test_data_batch_conversion",
+    size = "small",
+    srcs = ["tests/test_data_batch_conversion.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
     name = "test_deduping_schema",
     size = "small",
     srcs = ["tests/test_deduping_schema.py"],
@@ -911,6 +941,20 @@ py_test(
 )
 
 py_test(
+    name = "test_tensor_extension",
+    size = "small",
+    srcs = ["tests/test_tensor_extension.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
     name = "test_text",
     size = "small",
     srcs = ["tests/test_text.py"],
@@ -957,6 +1001,20 @@ py_test(
     name = "test_torch",
     size = "small",
     srcs = ["tests/test_torch.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_torch_tensor_utils",
+    size = "small",
+    srcs = ["tests/test_torch_tensor_utils.py"],
     tags = [
         "exclusive",
         "team:data",
@@ -1126,6 +1184,20 @@ py_test(
     name = "test_metadata_provider",
     size = "small",
     srcs = ["tests/test_metadata_provider.py"],
+    tags = [
+        "exclusive",
+        "team:data",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
+    name = "test_object_extension",
+    size = "small",
+    srcs = ["tests/test_object_extension.py"],
     tags = [
         "exclusive",
         "team:data",
