--- python/ray/train/util/__init__.py
@@ -1,7 +1,0 @@
-from ray.air.util.check_ingest import DummyTrainer
-
-__all__ = [
-    "DummyTrainer",
-]
-
-DummyTrainer.__module__ = "ray.train.util"
