+++ python/ray/train/v2/BUILD.bazel
@@ -70,6 +70,22 @@ py_test(
 )
 
 py_test(
+    name = "test_config",
+    size = "small",
+    srcs = ["tests/test_config.py"],
+    env = {"RAY_TRAIN_V2_ENABLED": "1"},
+    tags = [
+        "exclusive",
+        "team:ml",
+        "train_v2",
+    ],
+    deps = [
+        ":conftest",
+        "//:ray_lib",
+    ],
+)
+
+py_test(
     name = "test_circular_imports",
     size = "small",
     srcs = ["tests/test_circular_imports.py"],
